requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cattrs==25.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
flatbuffers==25.2.10
fonttools==4.58.4
gotrue==2.12.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jh2==5.0.9
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
niquests==3.14.1
numpy==2.3.1
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.1
packaging==25.0
pandas==2.3.0
pika==1.3.2
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
postgrest==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
qh3==1.5.3
realtime==2.5.3
requests==2.32.4
requests-cache==1.2.1
retry-requests==2.0.0
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.5.0
urllib3-future==2.13.900
wassima==1.2.2
websockets==15.0.1
xgboost==3.0.2
